cant eat coins

+++ src/com/company/Player.java
@@ -108,7 +108,9 @@ public class Player {
     public void eatFood(String food){
          if (findItemInBackpack(food)!=null) {
             Item playerItem = findItemInBackpack(food);
-            deleteItem(playerItem);
+            if(playerItem.equals(food)) {
+                deleteItem(playerItem);
+            }
 
         }else {
             System.out.println("cant find "+food);
